requirements.txt
asgiref==3.8.1
chardet==5.2.0
cssselect2==0.7.0
Django==4.2.11
django-admin-sortable2==2.1.10
django-filer==3.1.1
django-polymorphic==3.1.0
easy-thumbnails==2.8.5
lxml==5.2.1
mysqlclient==2.2.4
pillow==10.3.0
PyMySQL==1.1.0
python-dotenv==1.0.1
reportlab==4.2.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.11.0
webencodings==0.5.1
